+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+### Added
+
+- Added 'Run Detached' (run task in subprocess).  Uses [ConsoleControl](https://github.com/dwmkerr/consolecontrol).
+
 ## [7.0.13] - 2022-05-30
 
 ### Changed